+++ src/Resources/app/administration/src/decorator/rule-condition-service-decoration.js
@@ -1,0 +1,11 @@
+import '../core/component/is-shopgate';
+
+Shopware.Application.addServiceProviderDecorator('ruleConditionDataProviderService', (ruleConditionService) => {
+    ruleConditionService.addCondition('is_shopgate', {
+        component: 'is-shopgate',
+        label: 'Is Shopgate Mobile',
+        scopes: ['global']
+    });
+
+    return ruleConditionService;
+});