+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - added cart support with shipping, items (partial), promos, customer - check_cart
 - added order table
 - added generic shopgate payment method
+- added is_shopgate rule (untested)
 
 # 0.0.4
 - Added authentication configuration